+++ lib/punch/entities.rb
@@ -1,1 +1,0 @@
-require_relative 'entities/model'