+++ lib/punch/values.rb
@@ -1,0 +1,2 @@
+require_relative 'values/param'
+require_relative 'values/model'