+++ CHANGELOG.md
@@ -2,11 +2,20 @@
 
 > Solution does non exists unless tested
 
-- [ ] erb-helper tests parameters like `let(:args) {[args, **{}]}`
 - [ ] `<% include Punch::ErbHelper -%>` why this does not work?
 - [ ] content MD5 hashing; a file was "punched" but was it touched after?
 - [ ] `punch stat` prints some "punching" statistic based on MD5 and log
-- [ ] `punch` custom-erb by custom punchers, configure through punch.yml
+
+## [0.2.0] - 2022-07-31
+
+- Param and Model became Clean::Value
+- Param accepts default values like `param = nil`, `param:string = "42"`
+- Changed erb-helper for simplicity
+- Improved `exe/punch`;
+  - banner gives some examples;
+  - removed `punch sentry`.
+- Removed log info `Playbox changed to preview`
+- Improved README
 
 ## [0.1.1] - 2022-07-28